Split plugin lifecycle handling into a mixin.
